Create Arduino_Nano_328P
Arduino goes to sleep after 5 seconds of no detected pulses. When pulse detected the Arduino wakes up and also counts that pulse. All pulses are output to Serial Monitor, as well as when Arduino is going into sleep mode and how many times this has happened.

On board LED indicates if in sleep mode (LED off) or not (LED on)